Implement search in Zeya web client.
Add a search function which works by filtering the collection while it
is displayed in the page, looking for a search string in the title,
artist and album fields of each row.

A new search box is added in the top right corner of the page.

Split library loading and table rendering in two separate functions.

Reimplement next/previous movement functions so that they use the DOM to
walk the table instead of assuming that index values are consecutive
(which is not the case if the collection has been filtered).

New key binding: / to focus the search box.

Tested in Iceweasel 3.5.3 and Google Chrome 3.0.195.27.