zeya.py: enable traffic shaping even when buffered=true.
Traffic shaping only governs the rate of data transmission after the output
starts to get written (and after Content-Length is computed). Since disabling
traffic shaping doesn't allow the initial Content-Length header to get written
any faster (which is the major source of latency for WebKit clients), we might
as well turn on traffic shaping so that we don't also hose the connection.